260831_000002
컨트롤러.cs 초벌 작성

+++ Assets/01.Script/05.Character/CharacterController/CharacterStatus.cs
@@ -67,7 +67,7 @@ public class CharacterStatus
         DodgeRate = 0; // 상대의 명중보다 높으면 회피확률 발생
     }
 
-    public void AddExp(int amount)
+    public void AddExp(long amount)
     {
         Exp += amount;
     }
